feat(core): collapse plugins.py into server.py
Third microkernel cut. Resolves the PLAN-microkernel.md §1 / §6
contradiction Codex flagged: there is now no standalone plugins.py.
Every function and piece of module state from it lives in server.py.

Moved into server.py (near end, before if __name__):
  State:
    PLUGINS (Path), _LOCK, _lock_hashes, _DANGEROUS_PLUGINS,
    _cron_tasks, _start_time, IN_CONTAINER, _ENV_CEILING,
    _USER_MODE, MODE
  Lockfile:
    _load_lock(), _verify_plugin()
  Loaders:
    load_plugin() — Tier 0 disk loader
    unload_plugin() — generic, works for Tier 0 and 'lib:' Tier 1
    load_plugin_from_source() — Tier 1 source-string loader
    activate_lib_world() / deactivate_lib_world() — wrappers
    boot_load_active_lib() — iterate /lib/* at boot
    load_plugins() — Tier 0 disk scan + default install
  Audit/whitelist:
    _sync_actions_add() / _sync_actions_remove()
  Approval flow:
    handle_propose() / handle_approve() / register_plugin_routes()
  Async:
    cron_loop()

Inline changes in server.py:
  - Dropped `import plugins` and the `if plugins:` guard in the
    __main__ block (plugins module is unconditionally "us" now)
  - Removed `plugins.X(...)` prefixes — all direct calls
  - Kept the collision check (_find_lib_disk_collisions) as the
    boot gate before load_plugins() fires
  - `APPROVE_TOKEN` reconciled: server.py's existing env-only
    definition wins (plugins.py had a random-fallback that was
    effectively dead code since no auth path actually compared
    against it)
  - Inside load_plugin() + unload_plugin(), `server._auth = …`
    becomes `global _auth; _auth = …` since we're now in the
    defining module

plugins.py: DELETED.
plugins.lock: regenerated — server.py hash is the only change;
  `plugins/available/*.py` hashes unchanged.

Tests: 531 pass, 0 fail, 45 skip. No behaviour change; this is
pure relocation. The microkernel runtime surface is now:
  - server.py (core + Tier 0 disk loader + Tier 1 world driver)
  - plugins/available/*.py (Tier 0 plugin sources — admin is the
    last default; others are opt-in via /admin/load)
  - Tier 1 plugins live as /lib/<name> worlds

`example.py` and the `tools/migrate_disk_to_lib.py` transitional
script are not in this commit — they're the next two cuts of the
v4.5.0 breaking release.
