+++ plugins.lock
@@ -1,5 +1,4 @@
-c39d367899378cb7410c564367acc93d9b8ebd25260574487a7f51d9a70d6fd6  server.py
-aa876d0add9de918eb811dce6a3349096f15e0103ba29c8d2e5d98279200ff3b  plugins.py
+97ef75d99c412b074e1796f3069b25d2cd7f7bf15e8acdf9960e226ce454838f  server.py
 c101ad60d470cf6199dd064f673ca65f0e04ff732b0db3c21a6acd46bfc24254  plugins/available/admin.py
 bdf382c621d923502533b75bf61b8e1221ce42a1b795fc5820f5b0a7bfcdcf6c  plugins/available/backup.py
 1325b86979be69b6a3cd362093a9db3d4e4aec5aa38ceb0bfa2ed9bf5c00ae2a  plugins/available/browser.py
